Keep python API tests in one module
